refactor: split up the types file
